requirements.txt
httpx
fastapi
pydantic
requests
beautifulsoup4
uvicorn
anthropic
python-dotenv